+++ Changelog.md
@@ -3,6 +3,7 @@
 Features:
 
 Bugfixes:
+ * Code Generator: Allow ``block.blockhash`` without being called.
  * Code Generator: Properly skip unneeded storgae array cleanup when not reducing length.
  * Code Generator: Bugfix in modifier lookup in libraries.
  * Commandline interface: Support ``--evm-version constantinople`` properly.